requirements_simple.txt
# Essential packages for Mulberry Disease Prediction
# Core ML and Deep Learning
tensorflow>=2.15.0
keras>=3.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Jupyter Notebook Support
jupyter>=1.0.0
notebook>=6.5.0
ipykernel>=6.25.0

# Additional utilities
h5py>=3.9.0
